modify some expressions
