+++ test.py
@@ -1,0 +1,3 @@
+from blank_encode import *
+
+print(default_charset)
